requirements.txt
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
scipy>=1.11
plotly>=5.20
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
tabulate>=0.9
python-dotenv>=1.0
kaggle>=1.6
datasets>=2.18
huggingface_hub>=0.23
